+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 
 ### 🐞 Bug fixes
 
-- *...Add new stuff here...*
 - Dispose source resources on map style removal, it also fixes `cannot read properties of undefined (reading 'sourceCaches')` error (#1099).
+- Add MapGeoJSONFeature type as replacement for MapboxGeoJSONFeature. MapGeoJSONFeature type extends GeoJSONFeature type with layer, source, sourceLayer, and state properties (#1104).
 
 ## 2.1.8-pre.1